Moved the WapmContainer into its own module so we can abstract over WEBC v1
